Develop

+++ src/purchaserequest/purchaserequest.service.ts
@@ -20,6 +20,17 @@ export class PurchaseRequestService {
     postPurchaseRequestDto: PostPurchaseRequestDto,
     @Req() req: any,
   ) {
+    const bill = await this.prisma.billToPay.findFirst({
+      orderBy: { identifier: 'desc' },
+    });
+
+    let nextId: number;
+
+    if (bill?.identifier === null || bill?.identifier === undefined) {
+      nextId = 1;
+    } else {
+      nextId = bill.identifier + 1;
+    }
     const sizeArrayProductId = postPurchaseRequestDto.product_id.length;
 
     if (sizeArrayProductId === 0) {
@@ -30,6 +41,7 @@ export class PurchaseRequestService {
     }
 
     const data = {
+      identifier: nextId,
       reason: postPurchaseRequestDto.reason,
       status_id: '',
       active: postPurchaseRequestDto.active,
@@ -46,10 +58,12 @@ export class PurchaseRequestService {
     });
 
     try {
-      const statusWaiting = await this.prisma.status.findFirst({
-        where: { name: EStatus.waiting },
-      });
-      data.status_id = statusWaiting.id;
+      const { id: status_id, name: status_name } =
+        await this.prisma.status.findFirst({
+          where: { name: EStatus.waiting },
+        });
+
+      data.status_id = status_id;
 
       const result = await this.prisma.purchaseRequest.create({ data });
 
@@ -67,7 +81,6 @@ export class PurchaseRequestService {
               id: result.id,
             },
           });
-          console.log(ex);
         }
       }
 
@@ -84,6 +97,7 @@ export class PurchaseRequestService {
               where: {
                 group_id: managerGroups[g].id,
                 company_id: userCreateReq.company_id,
+                is_responsible: true,
               },
             });
 
@@ -91,7 +105,7 @@ export class PurchaseRequestService {
               await this.sendEmail(
                 userGestor.email,
                 process.env.FROM_EMAIL,
-                `Nova Solicitação de compra para aprovação (${statusWaiting.name})`,
+                `Nova Solicitação de compra para aprovação (${status_name})`,
                 `Olá ${userGestor.name}, há uma nova solicitação de compra criada pelo(a) ${userCreateReq.name}, aguardando aprovação. Para visualizar acesse: https://sistema.rta.eng.br`,
                 `<strong>Olá ${userGestor.name}, há uma nova solicitação de compra criada pelo(a) ${userCreateReq.name}, aguardando aprovação. Para visualizar acesse: https://sistema.rta.eng.br</strong><br><br><br><br>
                 Obs: Favor não responder este e-mail`,
@@ -106,7 +120,6 @@ export class PurchaseRequestService {
         message: `A Solicitação de compra foi criada com sucesso`,
       };
     } catch (ex) {
-      console.log(ex);
       return {
         status: false,
         message: `Não foi possível criar uma nova solicitação`,
@@ -129,24 +142,23 @@ export class PurchaseRequestService {
       };
     }
 
-    const purchaseRequestApproved =
-      await this.prisma.purchaseRequest.findUnique({
-        where: {
-          id,
-        },
-        include: {
-          Status: {
-            select: {
-              name: true,
-            },
+    const { Status } = await this.prisma.purchaseRequest.findUnique({
+      where: {
+        id,
+      },
+      include: {
+        Status: {
+          select: {
+            name: true,
           },
         },
-      });
+      },
+    });
 
-    if (purchaseRequestApproved.Status.name === EStatus.approvedDiretor) {
+    if (Status.name === EStatus.approvedDiretor) {
       return {
         status: false,
-        message: `Essa solicitação de compra se encontra com o status de APROVADA, não podendo ser alterada`,
+        message: `A solicitação encontra-se aprovada, não sendo possível realizar alteração`,
       };
     }
 
@@ -167,6 +179,20 @@ export class PurchaseRequestService {
       });
 
       try {
+        const { has_budget } = await this.prisma.purchaseRequest.findFirst({
+          where: {
+            id,
+          },
+        });
+
+        if (has_budget === true) {
+          return {
+            status: false,
+            message:
+              'O produto não pode ser excluído, pois já existe orçamento lançado para esta ordem.',
+          };
+        }
+
         await this.prisma.purchaseRequestProduct.deleteMany({
           where: {
             purchaserequest_id: id,
@@ -228,18 +254,26 @@ export class PurchaseRequestService {
       };
     }
 
-    const purchaseRequest = await this.prisma.purchaseRequest.findUnique({
+    const { Status } = await this.prisma.purchaseRequest.findUnique({
       where: { id },
+      include: {
+        Status: {
+          select: {
+            id: true,
+            name: true,
+          },
+        },
+      },
     });
 
-    const statusPurchaseRequest = await this.prisma.status.findFirst({
-      where: { id: purchaseRequest.status_id },
-    });
+    // const statusPurchaseRequest = await this.prisma.status.findFirst({
+    //   where: { id: purchaseRequest.status_id },
+    // });
 
-    if (statusPurchaseRequest.name === EStatus.approvedDiretor) {
+    if (Status.name === EStatus.approvedDiretor) {
       return {
         status: false,
-        message: `A solicitação de compra já se encontra aprovada, não sendo possível realizar alteração`,
+        message: `A solicitação encontra-se aprovada, não sendo possível realizar alteração`,
       };
     }
 
@@ -257,7 +291,7 @@ export class PurchaseRequestService {
           comment: patchPurchaseRequestDto.comment,
           approvedgestor_at: new Date(),
           approvedgestor_by: req.user.id,
-          is_approved_gestor: patchPurchaseRequestDto.is_approved,
+          is_approved_gestor: true,
         },
       };
 
@@ -302,7 +336,7 @@ export class PurchaseRequestService {
 
       return {
         status: true,
-        message: `A solicitação de compra foi aprovada pelo Gestor, agora precisa da aprovação do Diretor.`,
+        message: `A solicitação de compra foi aprovada pelo Gestor, necessitando da aprovação do Diretor.`,
       };
     }
   }
@@ -339,6 +373,16 @@ export class PurchaseRequestService {
       },
     });
 
+    if (
+      purchaseRequest.is_approved_gestor === false ||
+      purchaseRequest.is_approved_gestor === null
+    ) {
+      return {
+        status: false,
+        message: `A solicitação não foi aprovada pelo gestor.`,
+      };
+    }
+
     if (purchaseRequest.is_approved_gestor === true) {
       const statusApprovedDirector = await this.prisma.status.findFirst({
         where: {
@@ -353,21 +397,11 @@ export class PurchaseRequestService {
           comment: patchPurchaseRequestDto.comment,
           approveddiretor_at: new Date(),
           approveddiretor_by: req.user.id,
-          is_approved_diretor: patchPurchaseRequestDto.is_approved,
+          is_approved_diretor: true,
         },
       };
 
-      const diretor = await this.prisma.user.findUnique({
-        where: {
-          id: update.data.approveddiretor_by,
-        },
-      });
-
-      const status = await this.prisma.status.findFirst({
-        where: {
-          id: update.data.status_id,
-        },
-      });
+      await this.prisma.purchaseRequest.update(update);
 
       if (process.env.NODE_ENV === 'production') {
         const administrativoGroup = await this.prisma.group.findFirst({
@@ -381,31 +415,26 @@ export class PurchaseRequestService {
             where: {
               group_id: administrativoGroup.id,
               company_id: purchaseRequest.company_id,
+              is_responsible: true,
             },
           });
 
           await this.sendEmail(
             userAdministrativo.email,
             process.env.FROM_EMAIL,
-            `Nova Solicitação de compra para aprovação (${status.name})`,
-            `Olá ${userAdministrativo.name}, há uma nova solicitação de compra aprovada pelo(a) ${diretor.name}, aprovada com sucesso. Para visualizar acesse: https://sistema.rta.eng.br`,
-            `<strong>Olá ${userAdministrativo.name}, há uma nova solicitação de compra aprovada pelo(a) ${diretor.name}, aprovada com sucesso. Para visualizar acesse: https://sistema.rta.eng.br</strong><br><br><br><br>
+            `Nova Solicitação de compra para aprovação (${statusApprovedDirector.name})`,
+            `Olá ${userAdministrativo.name}, há uma nova solicitação de compra aprovada pelo(a) ${user.name}, aprovada com sucesso. Para visualizar acesse: https://sistema.rta.eng.br`,
+            `<strong>Olá ${userAdministrativo.name}, há uma nova solicitação de compra aprovada pelo(a) ${user.name}, aprovada com sucesso. Para visualizar acesse: https://sistema.rta.eng.br</strong><br><br><br><br>
                 Obs: Favor não responder este e-mail`,
           );
-          await this.prisma.purchaseRequest.update(update);
+          // await this.prisma.purchaseRequest.update(update);
         }
       }
 
-      const findBudget = await this.prisma.purchaseRequestBudget.findFirst({
+      const findBudgets = await this.prisma.purchaseRequestBudget.findMany({
         where: {
-          id: req.query.id,
+          purchaserequest_id: id,
           to_be_approved: true,
-        },
-      });
-
-      const productName = await this.prisma.purchaseRequestProduct.findMany({
-        where: {
-          id: req.query.id,
         },
         include: {
           Product: {
@@ -416,39 +445,121 @@ export class PurchaseRequestService {
         },
       });
 
-      let nameNewBillToPay = '';
+      const comprasPorFornecedor = new Map();
 
-      if (productName.length > 0) {
-        for (let i = 0; i < productName.length; i++) {
-          if (i < 1) {
-            nameNewBillToPay = productName[i].Product.name;
-          } else {
-            nameNewBillToPay =
-              nameNewBillToPay + '/' + productName[i].Product.name;
-          }
+      for (const compra of findBudgets) {
+        const {
+          supplier_id,
+          product_id,
+          Product,
+          to_be_approved,
+          budget,
+          shipping_fee,
+          quantity,
+        } = compra;
+
+        if (!comprasPorFornecedor.has(supplier_id)) {
+          comprasPorFornecedor.set(supplier_id, {
+            fornecedor: supplier_id,
+            produtosAprovados: [],
+            produtosData: [], // Array para armazenar dados de produtos
+            totalValue: 0,
+          });
+        }
+
+        if (to_be_approved === true) {
+          const fornecedorData = comprasPorFornecedor.get(supplier_id);
+          fornecedorData.produtosAprovados.push(product_id);
+
+          fornecedorData.produtosData.push({
+            nomeProduto: Product.name,
+            product_id,
+            budget,
+            quantity,
+            shipping_fee,
+          });
+
+          fornecedorData.totalValue += budget * quantity + shipping_fee;
         }
       }
 
-      const data = {
-        name: nameNewBillToPay,
-        type: 'SC',
-        supplier_id: findBudget.supplier_id,
-        price_approved: findBudget.budget + findBudget.shipping_fee,
-        price_updated: findBudget.budget + findBudget.shipping_fee,
-        created_by: req.user.id,
-        bill_status: 'A',
-        payment_info: '',
-        comment: '',
-        invoice_attachment: '',
-        company_id: purchaseRequest.company_id,
-        purchaserequest_id: purchaseRequest.identifier,
-      };
+      for (const compra of comprasPorFornecedor.values()) {
+        const { fornecedor, produtosData, totalValue } = compra;
 
-      await this.prisma.billToPay.create({ data });
+        let nomeProduto = '';
+
+        if (produtosData.length > 0) {
+          for (let i = 0; i < produtosData.length; i++) {
+            if (i < 1) {
+              nomeProduto = produtosData[i].nomeProduto.trim();
+            } else {
+              nomeProduto =
+                nomeProduto + '/' + produtosData[i].nomeProduto.trim();
+            }
+          }
+        }
+
+        const order = await this.prisma.purchaseOrder.create({
+          data: {
+            supplier_id: fornecedor,
+            purchaserequest_id: purchaseRequest.id,
+            company_id: purchaseRequest.company_id,
+          },
+        });
+
+        const data = await this.prisma.billToPay.create({
+          data: {
+            name: nomeProduto,
+            type: 'SC',
+            supplier_id: fornecedor,
+            price_approved: totalValue,
+            price_updated: totalValue,
+            created_by: req.user.id,
+            bill_status: 'A',
+            payment_info: '',
+            comment: '',
+            invoice_attachment: '',
+            dda: false,
+            company_id: purchaseRequest.company_id,
+            purchaserequest_identifier: purchaseRequest.identifier,
+          },
+        });
+
+        if (data.identifier) {
+          await this.prisma.purchaseOrder.update({
+            where: {
+              id: order.id,
+            },
+            data: {
+              billtopay_id: data.identifier,
+            },
+          });
+        }
+
+        for (const produtoData of produtosData) {
+          await this.prisma.purchaseOrderProduct.create({
+            data: {
+              quantity: produtoData.quantity,
+              price: produtoData.budget,
+              shipping_fee: produtoData.shipping_fee,
+              purchaseorder_id: order.id,
+              product_id: produtoData.product_id,
+            },
+          });
+
+          await this.prisma.productPrice.create({
+            data: {
+              product_id: produtoData.product_id,
+              price: produtoData.budget,
+              created_by: req.user.id,
+            },
+          });
+        }
+      }
 
       return {
         status: true,
-        message: `A solicitação de compra, foi aprovado com sucesso. Acesse contas a pagar para terminar de editar a nova conta criada.`,
+        message: `A solicitação de compra foi aprovada com sucesso. Acesse contas a pagar para terminar de editar a nova conta criada.`,
       };
     }
   }
@@ -505,6 +616,7 @@ export class PurchaseRequestService {
           where: {
             group_id: administrativoGroup.id,
             company_id: purchaseRequest.company_id,
+            is_responsible: true,
           },
         });
 
@@ -689,16 +801,30 @@ export class PurchaseRequestService {
       },
     });
 
+    if (statusName === 'TODOS') {
+      status.id = null;
+    }
+
+    let where = {};
+
+    if (!status.id) {
+      where = {
+        active,
+      };
+    } else {
+      where = {
+        active,
+        status_id: status.id,
+      };
+    }
+
     const purchaseRequest = await this.prisma.purchaseRequest.findMany({
-      take: 10,
-      skip: 10 * (page - 1),
+      take: 9,
+      skip: 9 * (page - 1),
       orderBy: {
         identifier: 'desc',
       },
-      where: {
-        active,
-        status_id: status.id,
-      },
+      where,
       include: {
         Status: {
           select: {
@@ -729,8 +855,25 @@ export class PurchaseRequestService {
       },
     });
 
+    if (name === 'TODOS') {
+      status.id = null;
+    }
+
+    let where = {};
+
+    if (!status.id) {
+      where = {
+        active,
+      };
+    } else {
+      where = {
+        active,
+        status_id: status.id,
+      };
+    }
+
     return await this.prisma.purchaseRequest.count({
-      where: { active, status_id: status.id },
+      where,
     });
   }
